[refactor][sdk] Change parameters to input arguments

+++ java/dingo-java-sdk-example/src/main/java/DingoJavaExampleUsingSDK.java
@@ -16,53 +16,70 @@ import java.util.Map;
 public class DingoJavaExampleUsingSDK {
 
     public static void main(String[] args) {
-
-        DingoClient dingoClient = new DingoClient("192.168.1.201:22001", 10);
-        dingoClient.open();
-
-        String tableName = "testChannel";
-
-        try {
-            boolean test1 = dingoClient.dropTable(tableName);
-            System.out.println(test1);
-        } catch (Exception e) {
-            System.out.println(e.toString());
+        if (args.length < 4) {
+            System.out.println("java -cp dingo-java-sdk-example-0.6.0-SNAPSHOT.jar DingoJavaExampleUsingSDK 192.168.1.201:22001 huzx 1000000000 true xxxxxxxxxx");
+            System.out.println("\t=> Args1: CoordinatorConnection(IP:PORT)");
+            System.out.println("\t=> Args2: TableName(ABC)");
+            System.out.println("\t=> Args3: RecordCnt(1000)");
+            System.out.println("\t=> Args4: IsReCreateTable(true: will drop table and create; else false)");
+            System.out.println("\t=> Args5: KeyPrefix(like xxxxx)");
+            System.exit(-1);
+        }
+        String coordinatorAddress = args[0];
+        String tableName = args[1];
+        Long  recordCnt = Long.parseLong(args[2]);
+        boolean isReCreateTable = args[3].equals("true") ? true : false;
+        String keyPrefix = "";
+        if (args.length >= 5) {
+            keyPrefix = args[4];
         }
 
-        ColumnDefinition c1 = new ColumnDefinition("id", "integer", "", 0, 0, false, 0, "");
+        System.out.println("1=>" + coordinatorAddress + ",2=>" + tableName + ",3=>" + recordCnt + ",4=>" + isReCreateTable);
+
+        DingoClient dingoClient = new DingoClient(coordinatorAddress, 10);
+        dingoClient.open();
+
+        ColumnDefinition c1 = new ColumnDefinition("id", "varchar", "", 0, 0, false, 0, "");
         ColumnDefinition c2 = new ColumnDefinition("name", "varchar", "", 0, 0, false, -1, "");
 
-        PartitionDetailDefinition detailDefinition = new PartitionDetailDefinition(null, null, Arrays.asList(new Object[]{1}));
+        PartitionDetailDefinition detailDefinition = new PartitionDetailDefinition(null, null, Arrays.asList(new Object[]{"1"}));
         PartitionRule partitionRule = new PartitionRule(null, null, Arrays.asList(detailDefinition));
 
         TableDefinition tableDefinition = new TableDefinition(tableName,
                 Arrays.asList(c1, c2),
                 1,
                 0,
-                partitionRule,
+                null,
                 Common.Engine.ENG_ROCKSDB.name(),
                 null);
-        boolean isSuccess = dingoClient.createTable(tableDefinition);
 
-        System.out.println("create table, result is " + isSuccess);
-
-        try {
-            Thread.sleep(12000);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
+        if (isReCreateTable) {
+            try {
+                dingoClient.dropTable(tableName);
+            } catch(Exception ex) {
+            }
+            boolean isSuccess = dingoClient.createTable(tableDefinition);
+            System.out.println("drop table and create table:" + isSuccess);
+            try {
+                Thread.sleep(12000);
+            } catch (InterruptedException e) {
+                throw new RuntimeException(e);
+            }
         }
 
-        int test_count = 100;
-        for (int i = 0; i < test_count; i++) {
+        Long test_count = recordCnt;
+        for (Long i = 0L; i < test_count; i++) {
             Map<String, Object> map = Maps.newLinkedHashMap();
-            map.put(c1.getName(), i);
-            map.put(c2.getName(), "zhangsan" + i);
+            map.put(c1.getName(), keyPrefix + i);
+            map.put(c2.getName(), keyPrefix + "=zhangsan=>" + i);
             boolean test = dingoClient.upsert(tableName, Collections.singletonList(new Record(c1.getName(), map)));
-            System.out.println("put " + i + " result is " + test);
+            if (i % 1000 == 0) {
+                System.out.println("Write key: " + i);
+            }
         }
 
-        for (int i = 0; i < test_count; i++) {
-            Record record = dingoClient.get(tableName, new Key(Arrays.asList(Value.get(i), Value.get(""))));
+        for (Long i = 0L; i < test_count; i++) {
+            Record record = dingoClient.get(tableName, new Key(Arrays.asList(Value.get(keyPrefix + i), Value.get(""))));
             System.out.println(record);
         }
